requirements.txt
numpy==1.13.1
opencv-python==3.3.0.10
tensorflow-gpu==1.5.0
tensorflow==1.5.0
natsort
Pillow
xml-python
pandas
regex